style(main): update page count formatting for consistency
- Changed the type casting of `page_count` in `main.cpp` from `unsigned long` to `uint32_t` for improved consistency and clarity in logging.
- Added a new function `net_publish_debug_ui` in `net.h` to publish a UI debug snapshot, enhancing the debugging capabilities of the firmware.
- Continued efforts to maintain a clean and consistent code style, contributing to overall code clarity and maintainability.

+++ firmware/arduino/src/main.cpp
@@ -1441,7 +1441,7 @@ static void smoke_full_window_test() {
     display.print("x");
     display.print(display.height());
   } while (display.nextPage());
-  Serial.printf("Pages drawn: %lu\n", (unsigned long)page_count);
+  Serial.printf("Pages drawn: %u\n", static_cast<uint32_t>(page_count));
 }
 
 // Removed minimal debug full_refresh
